TwoFactor: move constraint and validator to existing classes
